# Faster import part 3