--- a/aiida/cmdline/commands/cmd_devel.py
+++ b/aiida/cmdline/commands/cmd_devel.py
@@ -67,7 +67,7 @@ def devel_check_undesired_imports():
 
     for modulename in [
         'asyncio', 'requests', 'plumpy', 'disk_objectstore', 'paramiko', 'seekpath', 'CifFile', 'ase', 'pymatgen',
-        'spglib', 'pymysql'
+        'spglib', 'pymysql', 'yaml',
     ]:
         if modulename in sys.modules:
             echo.echo_warning(f'Detected loaded module "{modulename}"')