--- a/aiida/cmdline/commands/cmd_devel.py
+++ b/aiida/cmdline/commands/cmd_devel.py
@@ -66,8 +66,18 @@ def devel_check_undesired_imports():
     loaded_modules = 0
 
     for modulename in [
-        'asyncio', 'requests', 'plumpy', 'disk_objectstore', 'paramiko', 'seekpath', 'CifFile', 'ase', 'pymatgen',
-        'spglib', 'pymysql', 'yaml',
+        'asyncio',
+        'requests',
+        'plumpy',
+        'disk_objectstore',
+        'paramiko',
+        'seekpath',
+        'CifFile',
+        'ase',
+        'pymatgen',
+        'spglib',
+        'pymysql',
+        'yaml',
     ]:
         if modulename in sys.modules:
             echo.echo_warning(f'Detected loaded module "{modulename}"')