--- a/aiida/engine/processes/builder.py
+++ b/aiida/engine/processes/builder.py
@@ -13,8 +13,6 @@ import json
 from typing import TYPE_CHECKING, Any, Type
 from uuid import uuid4
 
-import yaml
-
 from aiida.engine.processes.ports import PortNamespace
 from aiida.orm import Dict, Node
 from aiida.orm.nodes.data.base import BaseType
@@ -245,6 +243,8 @@ class ProcessBuilder(ProcessBuilderNamespace):  # pylint: disable=too-many-ances
 
     def _repr_pretty_(self, p, _) -> str:  # pylint: disable=invalid-name
         """Pretty representation for in the IPython console and notebooks."""
+        import yaml
+
         return p.text(
             f'Process class: {self._process_class.__name__}\n'
             f'Inputs:\n{yaml.safe_dump(json.JSONDecoder().decode(PrettyEncoder().encode(self)))}'