--- a/aiida/engine/processes/calcjobs/monitors.py
+++ b/aiida/engine/processes/calcjobs/monitors.py
@@ -13,7 +13,9 @@ from aiida.common.lang import type_check
 from aiida.common.log import AIIDA_LOGGER
 from aiida.orm import CalcJobNode, Dict
 from aiida.plugins import BaseFactory
-from aiida.transports import Transport
+
+if t.TYPE_CHECKING:
+    from aiida.transports import Transport
 
 LOGGER = AIIDA_LOGGER.getChild(__name__)
 
@@ -171,7 +173,7 @@ class CalcJobMonitors:
     def process(
         self,
         node: CalcJobNode,
-        transport: Transport,
+        transport: 'Transport',
     ) -> CalcJobMonitorResult | None:
         """Call all monitors in order and return the result as one returns anything other than ``None``.