--- a/aiida/manage/external/__init__.py
+++ b/aiida/manage/external/__init__.py
@@ -14,15 +14,11 @@
 # yapf: disable
 # pylint: disable=wildcard-import
 
-from .postgres import *
 from .rmq import *
 
 __all__ = (
     'BROKER_DEFAULTS',
-    'DEFAULT_DBINFO',
     'ManagementApiConnectionError',
-    'Postgres',
-    'PostgresConnectionMode',
     'RabbitmqManagementClient',
     'get_launch_queue_name',
     'get_message_exchange_name',