@sphuber 2023-09-06
> I have no idea why the docs build is failing though...

```
aiida/engine/daemon/execmanager.py:docstring of aiida.engine.daemon.execmanager.upload_calculation:: WARNING: py:class reference target not found: MappingType
```
You can just add `py:class MappinType` to the `docs/source/nitpick-exceptions`