@danielhollas 2023-09-06
Thanks @sphuber. That indeed fixed it, although I still don't quite understand why this new warning appeared in this PR specifically, since I was not touching the `MappingType`.