@sphuber 2023-09-06
> Thanks @sphuber. That indeed fixed it, although I still don't quite understand why this new warning appeared in this PR specifically, since I was not touching the `MappingType`.

These docstring reference errors are notoriously vague and hard to debug. Which is why we have the ignore list, otherwise we would be stuck forever since they crop up so often for unexplicable reason.

I think the reason for the error is because we use an alias `MappingType` in the doc string of `upload_calculation` instead of the actual import `typing.Mapping`. I think this causes Sphinx to fail to resolve it. Why this happens only on this PR, I am not sure. Maybe because it uses a cache and had a successful build for the module from a previous build, but since now the file is touched, it rebuilt it causing the error to show. Anyway, it is not important, so let's just ignore the warning.